Update hessian code
